requirements.txt
fastapi
uvicorn
requests
GitPython
psycopg2-binary